Create PHP DB Access.rtf
This shows example how to create a form in jQuery, package it and run it against the database in a PHP call. The PHP call returns the status of the call as well as the JSON string it encodes just for example. This JSON can be viewed with an extension in your web browser and checked for correctness. This JSON can be used to dynamically build tables of products.

Example, we can keep the current cart as a CSV and build the table view directly by calling that user's "user_currentCart" DB entry